Update OCDynamicView.h
Code for drawing the dynamical system is now factored out
and resides in a class, e.g. RandomShapes or BrownianMotion,
that implements the DynamicalSystem protocol.  OCDynamicView
no longer refers to the OCViewController class.  Instead,
OCViewController, or any view controller using an OCDynamicView
class that wishes to report on frame rate and frame count should
implement the OCDynamicViewReporter protocol.